backend/requirements.txt
Flask==2.2.2 
Werkzeug==2.2.2
flask-cors==3.0.10
flask-swagger-ui==3.36.0
psycopg2-binary==2.9.10
SQLAlchemy==1.4.25
flasgger==0.9.5
starlette==0.27.0
bcrypt==3.2.0
ultralytics
torch
Pillow
torchvision
inference-sdk==0.31.1
python-dotenv==1.0.1